Many enhancements in all modules

+++ classification/utils/callbacks.py
@@ -1,24 +1,30 @@
+"""Callbacks for the training loop."""
+
 import os
 from abc import ABCMeta, abstractmethod
-from enum import StrEnum
+from enum import StrEnum, unique
+from os import PathLike
 from pathlib import Path
 from time import strftime
 from typing import Any
 
-import pandas as pd
 import torch
 from dotenv import find_dotenv, load_dotenv
 from torch.nn import Module
 from torch.optim import Optimizer  # type: ignore
 from torch.optim.lr_scheduler import LRScheduler
 
-from .common import init_logger, remove_file, save_learning_curves
+from .common import History, RecordedStats, init_logger, remove_file
+from .visualize import save_learning_curves
 
 assert load_dotenv(find_dotenv()), "The .env file is missing!"
+
+torch.serialization.add_safe_globals([RecordedStats])
 
 logger = init_logger(os.getenv("LOGGER"))
 
 
+@unique
 class Callbacks(StrEnum):
     """Available callbacks for the training loop."""
 
@@ -26,24 +32,39 @@ class Callbacks(StrEnum):
     ModelCheckpoint = "ModelCheckpoint"
     LearningCurvesCheckpoint = "LearningCurvesCheckpoint"
 
+    @classmethod
+    def keys(cls) -> set[str]:
+        return set(key.value for key in cls)
+
 
 class EarlyStopping:
-    def __init__(self, patience: int = 10, min_delta: float = 0.0) -> None:
+    """Early stopping callback with the given monitoring value."""
+
+    def __init__(
+        self,
+        monitor_value: str | RecordedStats = RecordedStats.VAL_LOSS,
+        patience: int = 10,
+        min_delta: float = 0.0,
+    ) -> None:
+        self.monitor_value = RecordedStats(monitor_value)
         self.patience = int(patience)
         self.min_delta = float(min_delta)
         self.counter = 0
-        self.min_validation_loss = float("inf")
+        self.min_value = float("inf")
 
-    def __call__(self, validation_loss: float) -> bool:
-        return self.should_stop(validation_loss)
+    def __call__(self, history: History) -> bool:
+        return self.should_stop(history)
 
-    def should_stop(self, validation_loss: float) -> bool:
-        if validation_loss < self.min_validation_loss:
-            self.min_validation_loss = validation_loss
+    def should_stop(self, history: History) -> bool:
+        current_values = history[self.monitor_value]
+        current_value = 0.0 if not current_values else current_values[-1]
+
+        if current_value < self.min_value:
+            self.min_value = current_value
             self.counter = 0
             return False
 
-        if validation_loss > self.min_validation_loss + self.min_delta:
+        if current_value > self.min_value + self.min_delta:
             self.counter += 1
             if self.counter >= self.patience:
                 return True
@@ -52,17 +73,19 @@ class EarlyStopping:
 
 
 class Checkpoint(metaclass=ABCMeta):
+    """Abstract base class for saving and loading checkpoints."""
+
     def __init__(
         self,
         checkpoints_ext: str,
         checkpoints_type: str,
-        checkpoints_dir: str | Path,
+        checkpoints_dir: str | PathLike,
     ) -> None:
         self.checkpoints_ext = str(checkpoints_ext)
         self.checkpoints_type = str(checkpoints_type)
         self.checkpoints_dir = Path(checkpoints_dir)
-        self.run_dir: Path | None = None
-        self.latest_file: Path | None = None
+        self.run_dir: PathLike | None = None
+        self.latest_file: PathLike | None = None
 
     def __call__(self, *args: Any, **kwargs: Any) -> None:
         self.save(*args, **kwargs)
@@ -91,13 +114,15 @@ class Checkpoint(metaclass=ABCMeta):
 
 
 class ModelCheckpoint(Checkpoint):
+    """Model checkpoint callback for saving and loading model states."""
+
     def __init__(
         self,
         model: Module,
         optimizer: Optimizer,
         scheduler: LRScheduler,
         checkpoints_freq: int = 1,
-        checkpoints_dir: str | Path = "./checkpoints/",
+        checkpoints_dir: str | PathLike = "./checkpoints/",
     ) -> None:
         checkpoints_type = "states"
         checkpoints_ext = ".tar"
@@ -108,9 +133,9 @@ class ModelCheckpoint(Checkpoint):
         self.scheduler = scheduler
         self.checkpoints_freq = int(checkpoints_freq)
         self.latest_epoch: int = 0
-        self.history: dict[str, list[float]] = dict()
+        self.history: History = dict()
 
-    def save(self, history: dict[str, list[float]]) -> None:
+    def save(self, history: History) -> None:
         super().create_run_storage()
         if self.run_dir is None:
             raise RuntimeError("Run directory not created.")
@@ -130,7 +155,8 @@ class ModelCheckpoint(Checkpoint):
                 },
                 current_file,
             )
-            remove_file(self.latest_file)
+            if self.latest_file is not None:
+                remove_file(self.latest_file)
             self.latest_file = current_file
 
     def load(self, map_location: str | torch.device | None = None) -> None:
@@ -148,13 +174,15 @@ class ModelCheckpoint(Checkpoint):
 
 
 class LearningCurvesCheckpoint(Checkpoint):
+    """Learning curves checkpoint callback for saving learning curves plots."""
+
     def __init__(
         self,
         smooth_out: bool = True,
         window: int = 5,
         order: int = 2,
         checkpoints_freq: int = 1,
-        checkpoints_dir: str | Path = "./checkpoints/",
+        checkpoints_dir: str | PathLike = "./checkpoints/",
     ) -> None:
         checkpoints_type = "plots"
         checkpoints_ext = ".png"
@@ -165,7 +193,7 @@ class LearningCurvesCheckpoint(Checkpoint):
         self.order = int(order)
         self.checkpoints_freq = int(checkpoints_freq)
 
-    def save(self, history: dict[str, list[float]]) -> None:
+    def save(self, history: History) -> None:
         super().create_run_storage()
         if self.run_dir is None:
             raise RuntimeError("Run directory not created.")
@@ -175,9 +203,9 @@ class LearningCurvesCheckpoint(Checkpoint):
             plot_file = Path(f"learning_curves_epoch_{epoch:03d}").with_suffix(self.checkpoints_ext)
             current_file = self.run_dir / plot_file
 
-            data = pd.DataFrame(history).rename_axis("epoch").reset_index().assign(epoch=lambda x: x.epoch.add(1))
-            save_learning_curves(data, current_file, self.window, self.order)
+            save_learning_curves(history, current_file, self.window, self.order)
             logger.debug(f"Saving learning curves to {current_file!s}...")
 
-            remove_file(self.latest_file)
+            if self.latest_file is not None:
+                remove_file(self.latest_file)
             self.latest_file = current_file
